Move checksum function to dedicated location
